requirements.txt
absl-py==0.14.0
alabaster==0.7.12
alembic==1.6.5
argcomplete==1.12.3
argon2-cffi==21.1.0
ase==3.22.1
astunparse==1.6.3
attrs==21.2.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
bleach==4.1.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
cfgv==3.3.0
charset-normalizer==2.0.2
click==8.0.1
cliff==3.8.0
cmaes==0.8.2
cmd2==2.1.2
cogdl==0.4.0
colorama==0.4.4
colorlog==5.0.1
cycler==0.10.0
d2l==0.17.0
debugpy==1.4.3
decorator==4.4.2
defusedxml==0.7.1
dgl-cu101==0.6.1
dgl-cu110==0.8.0.post1
dglgo==0.0.1
distlib==0.3.2
docutils==0.17.1
emoji==1.4.1
entrypoints==0.3
filelock==3.0.12
flake8==3.9.2
gast==0.3.3
gensim==3.8.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googledrivedownloader==0.4
grave==0.0.3
greenlet==1.1.0
grpcio==1.40.0
h5py==2.10.0
huggingface-hub==0.0.12
identify==2.2.11
idna==3.2
imagesize==1.3.0
importlib-metadata==4.6.1
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isodate==0.6.1
isort==5.10.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.3
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kaggle==1.5.12
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
littleutils==0.2.2
llvmlite==0.36.0
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.2.0
mkl-service==2.3.0
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
nodeenv==1.6.0
notebook==6.4.4
numba==0.53.1
numpy==1.21.5
numpydoc==1.2.1
oauthlib==3.1.1
ogb==1.3.1
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
opt-einsum==3.3.0
optuna==2.4.0
outdated==0.2.1
packaging==21.0
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.2
patsy==0.5.1
pbr==5.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
platformdirs==2.0.2
pre-commit==2.13.0
prettytable==2.1.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.0
psutil==2.1.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.20
pydantic==1.9.0
pyDeprecate==0.3.2
pyflakes==2.3.1
Pygments==2.10.0
pyhocon==0.3.58
pyod==0.9.0
pyparsing==2.4.7
pyperclip==1.8.2
pyrsistent==0.18.0
python-dateutil==2.8.2
python-editor==1.0.4
python-slugify==6.1.2
pytorch-ignite==0.4.4
pytorchtools==0.0.2
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.1
rdflib==6.1.1
regex==2021.7.6
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.3
Send2Trash==1.8.0
sentencepiece==0.1.96
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sklearn==0.0
smart-open==5.1.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
Sphinx==4.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.21
statsmodels==0.12.2
stevedore==3.3.0
tabulate==0.8.9
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
texttable==1.6.4
threadpoolctl==2.1.0
tokenizers==0.10.3
toml==0.10.2
torch==1.7.1
torch-cluster @ file:///home/ypd-23-teacher-2/xiaoqing/GNNPro/torch_cluster-1.5.8%2Bcu110-cp37-cp37m-linux_x86_64.whl
torch-geometric==1.6.1
torch-scatter @ file:///home/ypd-23-teacher-2/xiaoqing/GNNPro/torch_scatter-latest%2Bcu110-cp37-cp37m-linux_x86_64.whl
torch-sparse @ file:///home/ypd-23-teacher-2/xiaoqing/GNNPro/torch_sparse-latest%2Bcu110-cp37-cp37m-linux_x86_64.whl
torch-spline-conv @ file:///home/ypd-23-teacher-2/xiaoqing/GNNPro/torch_spline_conv-latest%2Bcu110-cp37-cp37m-linux_x86_64.whl
torch-tb-profiler==0.3.1
torchaudio==0.7.0a0+a853dff
torchmetrics==0.7.3
torchvision==0.8.2
tornado==6.1
tqdm==4.61.1
traitlets==5.1.0
transformers==4.8.2
typer==0.4.1
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1622748266870/work
urllib3==1.26.6
virtualenv==20.6.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xgboost==1.4.2
zipp==3.5.0
